Fix stack guard for entrypointless coroutines
Moved the "schedule loop" coroutine creation to the thread proc of the worker,
so it would set the stack protector to the correct thread's stack.

Also moved the whole Pre Write Barrier function setting logic to the GC:
previously the part of it that initialized threads upon their creation resided
in the ThreadManager.

Issue: #I8MYDQ

Signed-off-by: Konstantin Grebenschikov <grebenschikov.konstantin@huawei.com>