Fixed segfault in IrBuilder in JIT
The main problem of this bug is datarace for GetClass. Both runtime functions GetRuntime()->GetMethodById(...) and
GetRuntime()->IsMethodIntrinsic(...) try to do GetClass to get pointer to Class. We may store is not loaded method
(nullptr) to variable "method_" and in "if" check is loaded method, but after that use "nullptr" from first load.

Solution of this problem
Get method only after check for intrinsic, in that case if check return true, we sure that class and method are
loaded in ClassLinker and will not be changed.

Issue: https://gitee.com/openharmony/arkcompiler_runtime_core/issues/IAYR6G

Signed-off-by: Mikhail Kaskov <kaskov.mikhail@huawei.com>