Support FastPath and Native mode for intrinsics
Additionally support Native mode for irtoc implementations
of intrinsics (introduce special FastPathPlus) mode for the
parser. The Native version will be generated along with the
FastPath version of the function and suffixed with "_native".
The runtime should detect the presense of such implementations
and used suffixed names in such a case.

Issue: IBQQT5
Change-Id: I378eb89761050e0dd7205aa6af45ff6f68149ddb
Signed-off-by: Aaron Simkin <aaron.simkin1@huawei-partners.com>