Ignore erroneous cts test related to regspiller
Issue: https://gitee.com/openharmony/arkcompiler_ets_frontend/issues/ICPWGT
Internal issue: #27655

With the indexable objects set lowering patch
a CTS test fails. The fail is not related to
the patch itself, but a bug in regspiller. Put
the test on the ignore list, and annotate with
the existing issue for this bug.

Test: build, runtime tests

Change-Id: I0878e37fdb7d3a7bdfdd6dd672853efcf2f88b8b
Signed-off-by: Martin Sajti <martin.sajti@h-lab.eu>