Make cache of one-byte char strings non-movable
When the pointers to the cached one-byte char strings are located in
a non-movable memory, we are able to pass the reference to such cache
to the JIT compiler as it is without the need to load it in a register
using a `%tr`-based offset.

Also, because the cache is non-movable anymore, the
`void EtsPlatformTypes::UpdateCachesVmRefs(const GCRootUpdater &updater) const`
method, a method for updating the reference to the cache itself during
the compaction phase of the GC, is not required.

Test: build

Issue:
https://gitcode.com/openharmony/arkcompiler_runtime_core/issues/7762

Signed-off-by: Howard Roark <howard.roark1@huawei-partners.com>